Dynamic formulas for design components (MountedComponentTemplates) can now reference their mounts, vehicles, hulls, and empires. Still need to handle facilities, though, as well as anything else that might need a special context.

+++ FrEee/Modding/Loaders/ComponentLoader.cs
@@ -136,7 +136,7 @@ namespace FrEee.Modding.Loaders
 						// TODO - populate damage types once we implement them
 						w.DamageType = new DamageType { Name = rec.Get<string>("Weapon Damage Type", c)};
 
-						w.ReloadRate = rec.Get<int>("Weapon Reload Rate", c);
+						w.ReloadRate = rec.Get<double>("Weapon Reload Rate", c);
 
 						var wdisptype = rec.Get<string>("Weapon Display Type", c);
 						var wdispname = rec.Get<string>("Weapon Display", c);
